# By Michael Jackson (6) and others
# Via Michael Jackson (3) and Michael Groeber (1)
* develop:
  updating with bug fix notes
  fixing bug in MXAFileInfor where we ended up with infinite recursion due to passing in an empty string into the 'filename' method.
  Fixing output values so XYZ start at 1 instead of Zero
  Adding a writer to output a file that can be used as an input into the Ricardo Lebensohn's 3D Simulation FFT codes at Los Alamos National Lab.
  Setting Version Number
  updating gbcd to use doubles instead of float
  Directories and files are now being added to the prebuilts list using the new recursive algorithm.
  The directories are now created recursively, but the pipeline items are not.
  Continued work on recursive algorithm
  Adding requested features
  In the process of writing a recursive function to display any structure under the Prebuilt section

+++ Source/DREAM3DLib/IOFilters/VisualizeGBCD.h
@@ -138,7 +138,7 @@ class DREAM3DLib_EXPORT VisualizeGBCD : public SurfaceMeshFilter
     std::vector<OrientationMath::Pointer> m_OrientationOps;
 
   unsigned int* m_CrystalStructures;
-    float* m_GBCD;
+    double* m_GBCD;
 
     /**
      * @brief This function writes a set of Axis coordinates to that are needed
